requirements_gpu.txt
numpy
pytest
tensorflow-gpu>=1.6
tabulate
tqdm
